Finalisation du crud de la gestion d'evenements

+++ events/deleteEvent.php
@@ -9,6 +9,17 @@ if (isset($_GET['id'])) {
     header('Location: index.php');
 }
 
+$envetJoined = $db->prepare('SELECT * FROM items_users WHERE id_item = :id');
+$envetJoined->bindParam(':id', $id);
+$envetJoined->execute();
+$envetJoined = $envetJoined->fetchAll();
+if (count($envetJoined) > 0) {
+    echo "Vous ne pouvez pas supprimer un évènement qui a au moin sun utilisateur inscrit<br>";
+    echo "<a href='show.php?id=" . $id . "'>Afficher les Partipants</a>";
+    die;
+}
+
+
 $stmt = $db->prepare('DELETE FROM items WHERE id=:id');
 $stmt->bindParam('id', $id);
 $stmt->execute();
